set flag to refresh topic page

+++ views/topic.py
@@ -32,8 +32,7 @@ topic = Blueprint('topic', __name__)
 def index(page=1):
     msg = request.args.get('msg', None)
     list_page = get_user_topics(g.current_user.id, page)
-    if page >1:
-        #TODO ugly if page more than 1, topic will not hit the cache
+    if page >1 and list_page.last_time != get_user_topics(g.current_user.id, 1).last_time:
         backend.delete('topic:list:%d:%d' % (g.current_user.id, page))
         list_page = get_user_topics(g.current_user.id, page)
     return render_template('topic.index.html', msg=msg, \
